Refactor 'Delete revocation point' unit tests
